+++ runtime/service/src/main/java/org/apache/polaris/service/events/PolarisEventFilterConfiguration.java
@@ -1,0 +1,54 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+package org.apache.polaris.service.events;
+
+import io.quarkus.runtime.annotations.StaticInitSafe;
+import io.smallrye.config.ConfigMapping;
+import io.smallrye.config.WithParentName;
+import java.util.Map;
+import java.util.Optional;
+
+@StaticInitSafe
+@ConfigMapping(prefix = "polaris.event-filter")
+public interface PolarisEventFilterConfiguration {
+
+  /** Configuration of each named event filter. */
+  @WithParentName
+  Map<String, FilterConfiguration> filters();
+
+  interface FilterConfiguration {
+
+    /** The event filter implementation type. Currently supported: {@code jakarta-el}. */
+    String type();
+
+    /**
+     * For Jakarta EL filters only. The expression that events must match. The expression has access
+     * to {@code type}, {@code metadata}, and {@code attributes}. Use {@code ==} for equality
+     * predicates. Only built-in attributes can be specified for now.
+     */
+    Optional<String> include();
+
+    /**
+     * For Jakarta EL filters only. The expression that events must not match. The expression has
+     * access to {@code type}, {@code metadata}, and {@code attributes}. Use {@code ==} for equality
+     * predicates. Only built-in attributes can be specified for now.
+     */
+    Optional<String> exclude();
+  }
+}